improve: split up table of contents logic
