+++ docs/src/components/DocLayout/index.tsx
@@ -20,7 +20,8 @@ import { BookmarkProvider } from "../../services/bookmarks";
 import DocNavigation from "../DocNavigation";
 import Breadcrumbs from "../Breadcrumbs";
 import ComponentStatus from "../ComponentStatus";
-import TableOfContents, { useTableOfContents } from "../TableOfContents";
+import TableOfContents from "../TableOfContents";
+import { useTableOfContents } from "../../services/table-of-contents";
 import Tabs, { TabObject } from "../Tabs";
 import ReactExample from "../ReactExample";
 import Footer from "../Footer";
